third_party/blink/web_tests/fast/events/clientXY-in-zoom-and-scroll-expected.txt

Base
PASS event.clientX is 100
PASS event.clientY is 100
PASS event.pageX is 100
PASS event.pageY is 100

Just zoomed
PASS event.clientX is 83
PASS event.clientY is 83
PASS event.pageX is 83
PASS event.pageX is 83

Just scrolled
PASS event.clientX is 100
PASS event.clientY is 100
PASS event.pageX is 150
PASS event.pageY is 150

Zoomed and scrolled
PASS event.clientX is 83
PASS event.clientY is 83
PASS event.pageX is 133
PASS event.pageY is 133

RTL and scrolled
PASS event.clientX is 100
PASS event.clientY is 100
PASS event.pageX is 50
PASS event.pageY is 150
PASS successfullyParsed is true

TEST COMPLETE
